Move PaymentMethodTypes into the components->paymentmethod package
COSDK-500
